requirements.txt
# not-just-dev
numpy
scipy
attrs
networkx
matplotlib

mlflow